requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-multipart>=0.0.9

# Azure SDKs
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
openai>=1.12.0

# PDF processing
PyMuPDF>=1.23.0

# Data validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0
tenacity>=8.2.0

# Streamlit UI
streamlit>=1.31.0
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
tiktoken>=0.5.0
